BIGINT -> INTEGERに変更しました

+++ src/main/resources/schema.sql
@@ -2,14 +2,14 @@ DROP SCHEMA IF EXISTS 給与;
 CREATE SCHEMA 給与;
 
 CREATE TABLE 給与.ユーザー (
-  ユーザーID  BIGINT PRIMARY KEY,
+  ユーザーID  INTEGER PRIMARY KEY,
   登録日時 TIMESTAMP NOT NULL
 );
 CREATE SEQUENCE 給与.ユーザーIDシーケンサー;
 
 CREATE TABLE 給与.ユーザー名 (
-  ユーザー名ID BIGINT PRIMARY KEY,
-  ユーザーID  BIGINT NOT NULL,
+  ユーザー名ID INTEGER PRIMARY KEY,
+  ユーザーID  INTEGER NOT NULL,
   ユーザー名 VARCHAR(40) NOT NULL,
   登録日時 TIMESTAMP NOT NULL
     ,FOREIGN KEY (ユーザーID) REFERENCES 給与.ユーザー(ユーザーID)
@@ -17,8 +17,8 @@ CREATE TABLE 給与.ユーザー名 (
 CREATE SEQUENCE 給与.ユーザー名IDシーケンサー;
 
 CREATE TABLE 給与.ユーザー名対応表 (
-  ユーザーID  BIGINT NOT NULL,
-  ユーザー名ID BIGINT NOT NULL,
+  ユーザーID  INTEGER NOT NULL,
+  ユーザー名ID INTEGER NOT NULL,
     PRIMARY KEY (ユーザーID, ユーザー名ID)
     ,FOREIGN KEY (ユーザーID) REFERENCES  給与.ユーザー(ユーザーID)
     ,FOREIGN KEY (ユーザー名ID) 
@@ -26,8 +26,8 @@ CREATE TABLE 給与.ユーザー名対応表 (
 );
 
 CREATE TABLE 給与.ユーザー電話番号 (
-  ユーザー電話番号ID BIGINT PRIMARY KEY,
-  ユーザーID  BIGINT NOT NULL,
+  ユーザー電話番号ID INTEGER PRIMARY KEY,
+  ユーザーID  INTEGER NOT NULL,
   電話番号 VARCHAR(13) NOT NULL,
   登録日時 TIMESTAMP NOT NULL
     ,FOREIGN KEY (ユーザーID) REFERENCES 給与.ユーザー(ユーザーID)
@@ -35,8 +35,8 @@ CREATE TABLE 給与.ユーザー電話番号 (
 CREATE SEQUENCE 給与.ユーザー電話番号IDシーケンサー;
 
 CREATE TABLE 給与.ユーザー電話番号対応表 (
-  ユーザーID  BIGINT NOT NULL,
-  ユーザー電話番号ID BIGINT NOT NULL,
+  ユーザーID  INTEGER NOT NULL,
+  ユーザー電話番号ID INTEGER NOT NULL,
     PRIMARY KEY (ユーザーID, ユーザー電話番号ID)
     ,FOREIGN KEY (ユーザーID) REFERENCES  給与.ユーザー(ユーザーID)
     ,FOREIGN KEY (ユーザー電話番号ID) 
@@ -45,8 +45,8 @@ CREATE TABLE 給与.ユーザー電話番号対応表 (
 
 
 CREATE TABLE 給与.ユーザーメールアドレス (
-  ユーザーメールアドレスID BIGINT PRIMARY KEY,
-  ユーザーID  BIGINT NOT NULL,
+  ユーザーメールアドレスID INTEGER PRIMARY KEY,
+  ユーザーID  INTEGER NOT NULL,
   メールアドレス VARCHAR(255) NOT NULL,
   登録日時 TIMESTAMP NOT NULL
     ,FOREIGN KEY (ユーザーID) REFERENCES 給与.ユーザー(ユーザーID)
@@ -54,8 +54,8 @@ CREATE TABLE 給与.ユーザーメールアドレス (
 CREATE SEQUENCE 給与.ユーザーメールアドレスIDシーケンサー;
 
 CREATE TABLE 給与.ユーザーメールアドレス対応表 (
-  ユーザーID  BIGINT NOT NULL,
-  ユーザーメールアドレスID BIGINT NOT NULL,
+  ユーザーID  INTEGER NOT NULL,
+  ユーザーメールアドレスID INTEGER NOT NULL,
     PRIMARY KEY (ユーザーID, ユーザーメールアドレスID)
     ,FOREIGN KEY (ユーザーID) REFERENCES  給与.ユーザー(ユーザーID)
     ,FOREIGN KEY (ユーザーメールアドレスID) 
@@ -63,14 +63,14 @@ CREATE TABLE 給与.ユーザーメールアドレス対応表 (
 );
 
 CREATE TABLE 給与.削除済みユーザー (
-  ユーザーID  BIGINT PRIMARY KEY,
+  ユーザーID  INTEGER PRIMARY KEY,
   登録日時 TIMESTAMP NOT NULL
     ,FOREIGN KEY (ユーザーID) REFERENCES 給与.ユーザー(ユーザーID)
 );
 
 CREATE TABLE 給与.就業時間 (
-  就業時間ID BIGINT PRIMARY KEY,
-  ユーザーID BIGINT NOT NULL,
+  就業時間ID INTEGER PRIMARY KEY,
+  ユーザーID INTEGER NOT NULL,
   就業日 DATE NOT NULL,
   開始時刻 TIME NOT NULL,
   終了時刻 TIME NOT NULL,
@@ -80,9 +80,9 @@ CREATE TABLE 給与.就業時間 (
 );
 
 CREATE TABLE 給与.就業時間対応表 (
-  ユーザーID BIGINT NOT NULL,
+  ユーザーID INTEGER NOT NULL,
   就業日 DATE NOT NULL,
-  就業時間ID BIGINT NOT NULL,
+  就業時間ID INTEGER NOT NULL,
     PRIMARY KEY (ユーザーID, 就業日, 就業時間ID)
     ,FOREIGN KEY (ユーザーID) REFERENCES  給与.ユーザー(ユーザーID)
     ,FOREIGN KEY (就業時間ID)
